requirements.txt
django
bugsnag
djangorestframework